recommendation_engine/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
redis>=4.3.0
psycopg2-binary>=2.9.0
mlxtend>=0.20.0
matplotlib>=3.5.0
seaborn>=0.11.0 